requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
requests
requests-oauthlib
python-dotenv
scikit-learn
openai==0.28.0
pandas